Odradjena logika za brisanje automobila i za search automobila

+++ AgentApp/AgentApp/src/main/java/com/team32/AgentApp/controller/AutomobilController.java
@@ -4,6 +4,7 @@ import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.team32.AgentApp.DTO.AutomobilDTO;
@@ -29,12 +31,14 @@ import com.team32.AgentApp.model.entitet.Automobil;
 import com.team32.AgentApp.model.entitet.CommonData;
 import com.team32.AgentApp.model.entitet.Komentar;
 import com.team32.AgentApp.model.entitet.Ocena;
+import com.team32.AgentApp.model.entitet.Oglas;
 import com.team32.AgentApp.model.entitet.User;
 import com.team32.AgentApp.model.tentitet.SlikaVozila;
 import com.team32.AgentApp.service.AutomobilService;
 import com.team32.AgentApp.service.CommonDataService;
 import com.team32.AgentApp.service.KomentarService;
 import com.team32.AgentApp.service.OcenaService;
+import com.team32.AgentApp.service.OglasService;
 import com.team32.AgentApp.service.SlikaVozilaService;
 import com.team32.AgentApp.service.UserService;
 
@@ -58,6 +62,9 @@ public class AutomobilController {
 	
 	@Autowired
 	private SlikaVozilaService slikaVozService;
+	
+	@Autowired
+	private OglasService oglasService;
 	
 	@RequestMapping(method=RequestMethod.GET, value="/automobil", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<List<AutomobilDTO>> getAllAutomobil(Principal principal){
@@ -108,31 +115,31 @@ public class AutomobilController {
 	
 	
 	//Ovo je za automobil details vise odradjeno...
-		@RequestMapping(method=RequestMethod.GET, value="/automobil/{id}/details", produces = MediaType.APPLICATION_JSON_VALUE)
-		public ResponseEntity<AutomobilDetailsDTO> getAutomobilDetails(Principal principal, @PathVariable("id") Long id) throws Exception{
-			
-			//Preuzima se user iz sesije koji je trenutno ulogovan
-			String username = principal.getName();
-			
-			AutomobilDTO autoDTO = automobilService.findOneWithDetails(id);
-			
-			List<ReviewDTO> reviews = new ArrayList<>();
-			//Kreiramo review = dobavljamo komentare i ocene vezane za njega
-			if(getAllReviewsByAutomobilId(id, username) != null) {
-				reviews = getAllReviewsByAutomobilId(id, username);
-			}
-			
-			//Dobavljamo slike koje su vezane za njega
-			SlikaVozila s = slikaVozService.getSlikaVozilaByAutomobilId(id);
-			SlikaVozilaDTO slike = new SlikaVozilaDTO(s);
-//				
-//					
+	@RequestMapping(method=RequestMethod.GET, value="/automobil/{id}/details", produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<AutomobilDetailsDTO> getAutomobilDetails(Principal principal, @PathVariable("id") Long id) throws Exception{
+		
+		//Preuzima se user iz sesije koji je trenutno ulogovan
+		String username = principal.getName();
+		
+		AutomobilDTO autoDTO = automobilService.findOneWithDetails(id);
+		
+		List<ReviewDTO> reviews = new ArrayList<>();
+		//Kreiramo review = dobavljamo komentare i ocene vezane za njega
+		if(getAllReviewsByAutomobilId(id, username) != null) {
+			reviews = getAllReviewsByAutomobilId(id, username);
+		}
+		
+		//Dobavljamo slike koje su vezane za njega
+		SlikaVozila s = slikaVozService.getSlikaVozilaByAutomobilId(id);
+		SlikaVozilaDTO slike = new SlikaVozilaDTO(s);
+
+		
 //			if(getSlikeVozila(id) != null) {
 //				slike = getSlikeVozila(id);
 //			}	
-				
-			return new ResponseEntity<>(new AutomobilDetailsDTO(autoDTO, reviews, slike), HttpStatus.OK);
-		}
+			
+		return new ResponseEntity<>(new AutomobilDetailsDTO(autoDTO, reviews, slike), HttpStatus.OK);
+	}
 	
 	
 	@RequestMapping(method=RequestMethod.GET, value="/automobilReview", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -155,12 +162,94 @@ public class AutomobilController {
 				reviews = getAllReviewsByAutomobilId(a.getId(), username);
 			}
 			
-			
 			automobilRewiewDTO.add(new AutomobilReviewDTO(autoDTO, reviews));
 			
 		}
 		return new ResponseEntity<>(automobilRewiewDTO, HttpStatus.OK);
 	}
+	
+	@RequestMapping(method=RequestMethod.GET, value="/search", produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<List<AutomobilDTO>> searchApartments(Principal principal,
+			@RequestParam(value="markaAut", required=false) Long markaAutId,
+			@RequestParam(value="modelAut", required=false) Long modelAutId,
+			@RequestParam(value="klasaAut", required=false) Long klasaAutId,
+			@RequestParam(value="tipMenjaca", required=false) Long tipMenjacaId,
+			@RequestParam(value="tipGoriva", required=false) Long tipGorivaId,
+			@RequestParam(value="brojSedZaDec", required=false) Integer brojSedZaDec,
+			@RequestParam(value="ColDmgWaiv", required=false) Boolean ColDmgWaiv,
+			@RequestParam(value="kilometraza", required=false) Float kilometraza
+			){
+
+    	//Obrisati na kraju
+    	System.out.println("markaAut " + markaAutId);
+    	System.out.println("modelAut: " + modelAutId);
+    	System.out.println("klasaAut: " + klasaAutId);
+    	System.out.println("tipMenjaca: " + tipMenjacaId);
+    	System.out.println("tipGoriva: " + tipGorivaId);
+    	System.out.println("brojSedZaDec: " + brojSedZaDec);
+		System.out.println("ColDmgWaiv: " + ColDmgWaiv);
+		System.out.println("kilometraza: " + kilometraza);
+
+		//Preuzima se user iz sesije koji je trenutno ulogovan
+		String username = principal.getName();
+		User loggedAgent = userService.findByUsername(username);
+		
+		List<Automobil> allAgentsAutomobil = automobilService.getAllAutomobilByAgent(loggedAgent.getId());
+
+
+		if(markaAutId != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+					.filter(a -> a.getMarkaAutomobilaId() == markaAutId)
+					.collect(Collectors.toList());
+		}
+		if(modelAutId != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+				.filter(a -> a.getModelAutomobilaId() == modelAutId)
+				.collect(Collectors.toList());
+		}
+		if(klasaAutId != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+				.filter(a -> a.getKlasaAutomobilaId() == klasaAutId)
+				.collect(Collectors.toList());
+		}
+		if(tipMenjacaId != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+				.filter(a -> a.getTipMenjacaId() == tipMenjacaId)
+				.collect(Collectors.toList());
+		}
+		if (tipGorivaId != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+					.filter(a -> a.getTipGorivaId() == tipGorivaId)
+					.collect(Collectors.toList());
+		}
+		if (brojSedZaDec != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+					.filter(a -> a.getBrojSedistaZaDecu() == brojSedZaDec)
+					.collect(Collectors.toList());
+		}
+		if (ColDmgWaiv != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+					.filter(a -> a.isCollisionDamageWaiver() == ColDmgWaiv)
+					.collect(Collectors.toList());
+		}
+		if (kilometraza != null) {
+			allAgentsAutomobil = allAgentsAutomobil.stream()
+					.filter(a -> a.getPredjenaKilometraza() <= kilometraza)
+					.collect(Collectors.toList());
+		}
+		
+		//Oni automobili koji su zadovoljili uslove se prebacuju u DTO
+		List<AutomobilDTO> automobilDTO = new ArrayList<>();
+		for(Automobil a : allAgentsAutomobil) {
+			
+			automobilDTO.add(automobilService.findOneWithDetails(a.getId()));
+			
+		}
+
+		return new ResponseEntity<>(automobilDTO, HttpStatus.OK);
+    }
+	
+	
 	
 	@RequestMapping(method=RequestMethod.GET, value="/automobil/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<AutomobilDTO> getAutomobil(@PathVariable("id") Long id){
@@ -174,45 +263,21 @@ public class AutomobilController {
 	}
 
 	@RequestMapping(value="/automobil/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Void> deleteAutomobil(@PathVariable Long id){
+	public ResponseEntity<Void> deleteAutomobil(@PathVariable Long id) throws Exception{
 		Automobil automobil = automobilService.findOne(id);
 		if(automobil != null) {
+			
+			//Provera da li postoji neki oglas kreiran za taj automobil
+			if(oglasService.getAllOglasByAutomobilId(automobil.getId()).size() != 0) {
+				throw new Exception("There is an ad connected to this car!");
+			}
+			
 			automobilService.deleteAutomobil(id);
 			comDataService.deleteCommonData(automobil.getCommonDataId());
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		}else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
-	}
-
-	
-	public List<ReviewDTO> getAllReviewsByAutomobilId(Long autoId, String loggedUsername){
-		List<ReviewDTO> reviews = new ArrayList<>();
-		
-		//Komentari i ocene se poklapaju po 2 kriterijuma po automobilu i po rezervaciji
-		//Vadimo sve komentare i rezervacije za jedan isti automobil;
-		List<Komentar> sviKomentAutomob = komentarService.getAllKomentarByAutomobilId(autoId);
-		List<Ocena> sveOceneAutomob = ocenaService.getAllOcenaByAutomobilId(autoId);
-		
-		//Prolazimo kroz rezulat i proveravamo koji od njih pripadaju istoj rezervaciji
-		for(Komentar k : sviKomentAutomob) {
-			//Provera da li je komentar ostavio sam agent (onda nema ocene)
-			CommonData comData = comDataService.findOne(k.getCommonDataId());
-			User u = userService.findOne(comData.getUserid());
-			
-			if(u.getKorisnickoIme().equals(loggedUsername)  && k.isOdobren() == true) {
-				reviews.add(new ReviewDTO(new KomentarDTO(k, u.getKorisnickoIme())));
-				continue;
-			}
-			for(Ocena o : sveOceneAutomob) {
-				if(k.getRezervacijaId() == o.getRezervacijaId() && k.isOdobren() == true) {
-					
-					reviews.add(new ReviewDTO(new OcenaDTO(o), new KomentarDTO(k, u.getKorisnickoIme())));
-				}
-			}
-		}
-		
-		return reviews;
 	}
 	
 	@RequestMapping(method=RequestMethod.POST, value="/automobil", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -291,6 +356,35 @@ public class AutomobilController {
 		return new ResponseEntity<>(auto,HttpStatus.OK);
 	}
 	
+	
+	public List<ReviewDTO> getAllReviewsByAutomobilId(Long autoId, String loggedUsername){
+		List<ReviewDTO> reviews = new ArrayList<>();
+		
+		//Komentari i ocene se poklapaju po 2 kriterijuma po automobilu i po rezervaciji
+		//Vadimo sve komentare i rezervacije za jedan isti automobil;
+		List<Komentar> sviKomentAutomob = komentarService.getAllKomentarByAutomobilId(autoId);
+		List<Ocena> sveOceneAutomob = ocenaService.getAllOcenaByAutomobilId(autoId);
+		
+		//Prolazimo kroz rezulat i proveravamo koji od njih pripadaju istoj rezervaciji
+		for(Komentar k : sviKomentAutomob) {
+			//Provera da li je komentar ostavio sam agent (onda nema ocene)
+			CommonData comData = comDataService.findOne(k.getCommonDataId());
+			User u = userService.findOne(comData.getUserid());
+			
+			if(u.getKorisnickoIme().equals(loggedUsername)  && k.isOdobren() == true) {
+				reviews.add(new ReviewDTO(new KomentarDTO(k, u.getKorisnickoIme())));
+				continue;
+			}
+			for(Ocena o : sveOceneAutomob) {
+				if(k.getRezervacijaId() == o.getRezervacijaId() && k.isOdobren() == true) {
+					
+					reviews.add(new ReviewDTO(new OcenaDTO(o), new KomentarDTO(k, u.getKorisnickoIme())));
+				}
+			}
+		}
+		
+		return reviews;
+	}
 	
 	//Pomocna metoda za kreiranje novog commonData
 	public CommonData setCommonData(Long userId) throws Exception {
